Add ui:history module & cleanup app module
