+++ app/build.gradle.kts
@@ -1,7 +1,6 @@
 plugins {
     id("com.android.application")
     kotlin("android")
-    kotlin("android.extensions")
     kotlin("kapt")
     id("dagger.hilt.android.plugin")
 }
@@ -38,7 +37,6 @@ android {
     }
 
     compileOptions {
-        isCoreLibraryDesugaringEnabled = true
         sourceCompatibility = JavaVersion.VERSION_1_8
         targetCompatibility = JavaVersion.VERSION_1_8
     }
@@ -67,8 +65,6 @@ android {
 }
 
 dependencies {
-    implementation(fileTree(mapOf("dir" to "libs", "include" to listOf("*.jar", "*.aar"))))
-
     implementation(project(":repo"))
     implementation(project(":common"))
     implementation(project(":scrobble"))
@@ -79,20 +75,15 @@ dependencies {
     implementation(project(":ui:profile"))
     implementation(project(":ui:detail"))
     implementation(project(":ui:search"))
+    implementation(project(":ui:history"))
 
     // AndroidX
-    implementation(AndroidX.browser)
-    implementation(AndroidX.core.ktx)
     implementation(AndroidX.appCompat)
-    implementation(AndroidX.activityKtx)
     implementation(AndroidX.compose.runtime.liveData)
     implementation("androidx.navigation:navigation-compose:_")
-    implementation(AndroidX.ui.tooling)
     implementation(AndroidX.lifecycle.liveDataKtx)
     implementation(AndroidX.lifecycle.viewModelKtx)
-    implementation(AndroidX.lifecycle.viewModelSavedState)
     implementation(AndroidX.work.runtimeKtx)
-    implementation(AndroidX.paletteKtx)
     implementation(AndroidX.hilt.lifecycleViewModel)
     implementation(AndroidX.hilt.work)
     kapt(AndroidX.hilt.compiler)
@@ -103,7 +94,6 @@ dependencies {
 
     // Other
     debugImplementation(Square.leakCanary.android)
-    coreLibraryDesugaring("com.android.tools", "desugar_jdk_libs", "_")
 
     // Testing
     testImplementation(Testing.junit4)
