+++ app/src/main/java/de/schnettler/scrobbler/MainScreen.kt
@@ -1,4 +1,4 @@
-package de.schnettler.scrobbler.screens
+package de.schnettler.scrobbler
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -8,6 +8,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import de.schnettler.database.models.LastFmEntity
 import de.schnettler.scrobbler.ui.charts.ChartScreen
+import de.schnettler.scrobbler.ui.charts.ChartsViewModel
 import de.schnettler.scrobbler.ui.common.compose.navigation.Screen
 import de.schnettler.scrobbler.ui.common.compose.navigation.UIAction
 import de.schnettler.scrobbler.ui.common.compose.navigation.UIError
@@ -15,22 +16,23 @@ import de.schnettler.scrobbler.ui.detail.DetailScreen
 import de.schnettler.scrobbler.ui.detail.viewmodel.AlbumViewModel
 import de.schnettler.scrobbler.ui.detail.viewmodel.ArtistViewModel
 import de.schnettler.scrobbler.ui.detail.viewmodel.TrackViewModel
+import de.schnettler.scrobbler.ui.history.HistoryScreen
+import de.schnettler.scrobbler.ui.history.LocalViewModel
 import de.schnettler.scrobbler.ui.profile.ProfileScreen
+import de.schnettler.scrobbler.ui.profile.UserViewModel
 import de.schnettler.scrobbler.ui.search.SearchScreen
 import de.schnettler.scrobbler.ui.search.SearchViewModel
 import de.schnettler.scrobbler.ui.settings.SettingsScreen
-import de.schnettler.scrobbler.util.SessionState
+import de.schnettler.scrobbler.model.SessionState
 import de.schnettler.scrobbler.util.destination
 import de.schnettler.scrobbler.util.secondOrNull
-import de.schnettler.scrobbler.viewmodels.LocalViewModel
-import de.schnettler.scrobbler.viewmodels.MainViewModel
 
 @Composable
 fun MainRouteContent(
     model: MainViewModel,
     navController: NavHostController,
-    chartsModel: de.schnettler.scrobbler.ui.charts.ChartsViewModel,
-    userViewModel: de.schnettler.scrobbler.ui.profile.UserViewModel,
+    chartsModel: ChartsViewModel,
+    userViewModel: UserViewModel,
     localViewModel: LocalViewModel,
     searchViewModel: SearchViewModel,
     artistViewModel: ArtistViewModel,
