--- app/src/main/java/de/schnettler/scrobbler/screens/Previews.kt
@@ -1,28 +1,0 @@
-package de.schnettler.scrobbler.screens
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.runtime.Composable
-import androidx.ui.tooling.preview.Preview
-import de.schnettler.database.models.LastFmEntity.Artist
-import de.schnettler.scrobbler.ui.common.compose.widget.MediaCard
-
-@Preview
-@Composable
-fun ListingCardPreview() {
-        Column {
-            val artist = Artist(
-                name = "Dreamcatcher",
-                url = "Url"
-            )
-            MediaCard(
-                name = artist.name,
-                plays = 10,
-                imageUrl = null
-            ) {}
-            MediaCard(
-                name = artist.name,
-                plays = -1,
-                imageUrl = null
-            ) {}
-        }
-}
