+++ app/src/main/java/de/schnettler/scrobbler/util/ListUtils.kt
@@ -1,0 +1,5 @@
+package de.schnettler.scrobbler.util
+
+fun <T> List<T>.secondOrNull(): T? {
+    return if (size < 2) null else this[1]
+}
