+++ ui/history/src/main/java/de/schnettler/scrobbler/ui/history/LocalViewModel.kt
@@ -1,4 +1,4 @@
-package de.schnettler.scrobbler.viewmodels
+package de.schnettler.scrobbler.ui.history
 
 import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.viewModelScope
